Fix multiple bugs in time based policies
The keepConsistent method had two main issues. One is that it failed to
properly clear all the buckets it should. The second is that it failed
if given a value representing the past.

The first issue of failing to clear all the buckets was already
addressed by a previous commit. That commit added logic to the append
method to detect bucket selection changes and reset the bucket at that
time. This resolved the issue but only for cases where a value was
appended before running a reduce because the keepConsistent call in the
reduce method was not guarded with this same logic. In this patch, I've
updated the keepConsistent method to correctly clear all relevant
buckets and removed the bucket reset logic from the append method. This
both passes the regression test added by the older commit and reduced
the number of allocations.

The second issue of handling past data was exposed by the
AppendWithTimestamp addition but was technically present before that
method was added. While the Go time.Now does include a monotonic
component on most platforms, all of our time based operations leverage
wall clock values. It's possible for wall clock values to go backwards
under normal conditions of a system and especially so if the timestamp
value is provided by the user instead of the system. To account for
this, the Append and AppendWithTimestamp methods now discard any values
that are earlier in the past than the current window represents. The
current window is always defined by the latest timestamp given to
append. Any timestamp representing a value later than the latest
timestamp value given to append shifts the window forward. Overall, the
system works exactly as before with respect to always increasing
timestamps but now can tolerate backwards time within the window.

This patch also makes some minor fixes to the time based tests. Notably,
there was a timing based race condition that caused intermittent test
failures in the TestTimeWindow. The issue was that the test loop would
append and then sleep one bucket duration such that the last element
inserted may or may not result in the first bucket being trucated when
calling Reduce. Hitting this case depending on jitter that added a delay
between the last sleep and the reduction. The test now sleeps after all
but the last append which better guarantees that the last bucket has not
rolled over before calling reduce. To account, the last sleep is moved
outside the loop and after the reduction in order to continue the test.