Modify PointPolicy to lazy establish points
This patch alters the PointPolicy to only add entries to buckets when a
point is actually added to the window. Previously, the policy would
eagerly create one point in each bucket with the zero value of the
window type. This resulted in unexpected behavior when, for example,
counting points in an empty point policy because it would show the
number of buckets rather than the number of written point.